Moved chans to pkg/chans
