+++ constraints.go
@@ -58,21 +58,3 @@ type Complex interface {
 type Ordered interface {
 	Integer | Float | ~string
 }
-
-// RecvChan is a constraint that permits a receive-only chan or a send & receive
-// channal.
-type RecvChan[T any] interface {
-	~chan T | ~<-chan T
-}
-
-// SendChan is a constraint that permits a send-only chan or a send & receive
-// channal.
-type SendChan[T any] interface {
-	~chan T | ~chan<- T
-}
-
-// Chan is a constraint that permits any type of channel, be it a receive-only,
-// send-only, or unidirectional channel.
-type Chan[T any] interface {
-	~chan T | ~chan<- T | ~<-chan T
-}
