+++ pkg/chans/chans.go
@@ -2,15 +2,17 @@
 //
 // SPDX-License-Identifier: MIT
 
-package typ
+package chans
 
 import (
 	"context"
 	"time"
+
+	"gopkg.in/typ.v3"
 )
 
 // SendTimeout sends a value to a channel, or cancels after a given duration.
-func SendTimeout[C SendChan[V], V any](ch C, value V, timeout time.Duration) bool {
+func SendTimeout[C Sender[V], V any](ch C, value V, timeout time.Duration) bool {
 	if timeout <= 0 {
 		ch <- value
 		return true
@@ -27,7 +29,7 @@ func SendTimeout[C SendChan[V], V any](ch C, value V, timeout time.Duration) boo
 
 // SendContext receives a value from a channel, or cancels when the given
 // context is cancelled.
-func SendContext[C SendChan[V], V any](ctx context.Context, ch C, value V) bool {
+func SendContext[C Sender[V], V any](ctx context.Context, ch C, value V) bool {
 	select {
 	case ch <- value:
 		return true
@@ -38,7 +40,7 @@ func SendContext[C SendChan[V], V any](ctx context.Context, ch C, value V) bool 
 
 // RecvTimeout receives a value from a channel, or cancels after a given timeout.
 // If the timeout duration is zero or negative, then no limit is used.
-func RecvTimeout[C RecvChan[V], V any](ch C, timeout time.Duration) (V, bool) {
+func RecvTimeout[C Receiver[V], V any](ch C, timeout time.Duration) (V, bool) {
 	if timeout <= 0 {
 		value, ok := <-ch
 		return value, ok
@@ -49,7 +51,7 @@ func RecvTimeout[C RecvChan[V], V any](ch C, timeout time.Duration) (V, bool) {
 		timer.Stop()
 		return value, ok
 	case <-timer.C:
-		return Zero[V](), false
+		return typ.Zero[V](), false
 	}
 }
 
@@ -60,14 +62,14 @@ func RecvContext[C ~<-chan V, V any](ctx context.Context, ch C) (V, bool) {
 	case value, ok := <-ch:
 		return value, ok
 	case <-ctx.Done():
-		return Zero[V](), false
+		return typ.Zero[V](), false
 	}
 }
 
 // RecvQueued will receive all values from a channel until either there's no
 // more values in the channel's queue buffer, or it has received maxValues
 // values, or until the channel is closed, whichever comes first.
-func RecvQueued[C RecvChan[V], V any](ch C, maxValues int) []V {
+func RecvQueued[C Receiver[V], V any](ch C, maxValues int) []V {
 	var buffer []V
 	for len(buffer) < maxValues {
 		select {
@@ -84,7 +86,7 @@ func RecvQueued[C RecvChan[V], V any](ch C, maxValues int) []V {
 // more values in the channel's queue buffer, or it has filled buf with
 // values, or until the channel is closed, whichever comes first, and then
 // returns the number of values that was received.
-func RecvQueuedFull[C RecvChan[V], B ~[]V, V any](ch C, buf B) int {
+func RecvQueuedFull[C Receiver[V], B ~[]V, V any](ch C, buf B) int {
 	var index int
 	for index < len(buf) {
 		select {
