+++ pkg/chans/publisher.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
-package typ
+package chans
 
 import (
 	"errors"
